added jwt claims

+++ 7-Token-refreshing-and-Flask-JWT-Extended/resources/item.py
@@ -1,3 +1,4 @@
+from flask_jwt_extended.utils import get_jwt
 from flask_restful import Resource, reqparse
 from flask_jwt_extended import jwt_required
 from models.item import ItemModel
@@ -38,7 +39,12 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin previleges required'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
